Bring up to date
Bring the project layout up to date to be more in line with our other
Serverless projects, making it easier to make changes to it in the
future.
