+++ .github/workflows/pr-tests.yml
@@ -1,17 +1,20 @@
 name: PR Tests
+
 on:
   pull_request:
+    paths-ignore:
+      - '**/*.md'
 
 jobs:
   test:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v2
-    - name: Set up Python 3.9
-      uses: actions/setup-python@v2
-      with:
-        python-version: 3.9
-    - name: Install dependencies
-      run: pip install tox
-    - name: Test with tox
-      run: tox -e py,flake8,black
+      - uses: actions/checkout@b4ffde65f46336ab88eb53be808477a3936bae11 # v4.1.1
+      - name: Set up Python 3.11
+        uses: actions/setup-python@0a5c61591373683505ea898e09a3ea4f39ef2b9c # v5.0.0
+        with:
+          python-version: '3.11'
+      - name: Install dependencies
+        run: pip install tox
+      - name: Test with tox
+        run: tox -e py,flake8,black
