+++ bydelsfakta_api/exceptions.py
@@ -1,0 +1,6 @@
+class IllegalFormatError(Exception):
+    pass
+
+
+class S3FileNotFoundError(Exception):
+    pass
