+++ tox.ini
@@ -1,17 +1,19 @@
 [tox]
-envlist = py39, flake8, black
+envlist = py311, flake8, black
 
 [testenv]
 deps=
+    moto
     pytest
-    moto==1.3.8
-    requests-mock==1.5.2
+    requests-mock
 commands=
     pytest
 setenv =
     AWS_ACCESS_KEY_ID = mock
     AWS_SECRET_ACCESS_KEY = mock
     AWS_XRAY_SDK_ENABLED = false
+    METADATA_API_URL = https://example.org/test
+    STAGE = test
 
 [testenv:flake8]
 skip_install = true
@@ -26,7 +28,6 @@ deps =
     black
 commands =
     black --check .
-
 
 [flake8]
 # https://github.com/ambv/black/blob/master/.flake8
